LayerMetadata returns layer blob size
We are interested in getting the size of an image prior to downloading it via
the registry API, so we would like to use this endpoint in order to do so
since it appears to be the only documented, supported way. If anyone knows
otherwise I would be just delighted to hear it. It appears at present we have
to get the manifest and get the size of each layer and sum those numbers to
get the total image size. This was slightly disappointing since
`(*SignedManifest).References()` appears to return a struct that contains the
size, but never appears to fill in size (nor does the manifest API return such
info) -- it is not the end of the world.

HasLayer calls the same endpoint as LayerMetadata but does not expose the size
through its API, hence the new method LayerMetadata. Any style preferences
welcome.

Thank you for building this and contributing it back to the community!